tired of thinking of stupid commit messages

+++ src/com/kport/langueg/parse/ast/nodes/expr/integer/NInt16.java
@@ -26,7 +26,7 @@ public class NInt16 extends NExpr {
 
     @Override
     protected String nToString() {
-        return Long.toString(val);
+        return Short.toString(val);
     }
 
     @Override
