+++ src/com/kport/langueg/parse/ast/nodes/expr/integer/NUInt32.java
@@ -26,7 +26,7 @@ public class NUInt32 extends NExpr {
 
     @Override
     protected String nToString() {
-        return Long.toUnsignedString(val);
+        return Integer.toUnsignedString(val);
     }
 
     @Override
